pyproject.toml
[tool.poetry]
name = "sparse-vibert-experiments"
version = "0.1.0"
description = "Reproducible experiments for Sparse-VIBERT paper"
authors = ["AI Researcher <research@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
# Deep-learning stack
torch = "^2.2.0"
transformers = "^4.38.0"
datasets = "^2.18.0"
# Experiment management
hydra-core = "^1.3.2"
wandb = "^0.16.4"
optuna = "^3.5.0"
# Analysis & visualisation
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
pandas = "^2.2.0"
scipy = "^1.12.0"
scikit-learn = "^1.4.0"
pyyaml = "^6.0.1"
tqdm = "^4.66.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"